+++ kernels/allocate_stringset.cuh
@@ -15,9 +15,7 @@
 //kernel outputs: seqsh, seqsv, lids
 // fill StringSet
 
-//void fill_stringset_cuda(uint64_t beg, uint64_t end, uint64_t local_nnz_count, char** seqsh_str, char** seqsv_str, uint64_t *lids, uint64_t* mattuples1, uint64_t* mattuples2, uint64_t* mattuples3, uint64_t row_offset, uint64_t col_offset, int ckthr, char** dfd_col_seq_gpu, char** dfd_row_seq_gpu, uint64_t* align_cts);
-
-void fill_stringset_cuda(uint64_t beg, uint64_t end,uint64_t local_nnz_count, char **seqsh_str, char **seqsv_str, uint64_t *lids, uint64_t *mattuples1, uint64_t *mattuples2, uint64_t *mattuples3, uint64_t row_offset, uint64_t col_offset, int ckthr, char **dfd_col_seq_gpu, char **dfd_row_seq_gpu, uint64_t* align_cnts);
+void fill_stringset_cuda(uint64_t beg, uint64_t end,uint64_t local_nnz_count, char **seqsh_str, char **seqsv_str, uint64_t *lids, int *mattuples1, int *mattuples2, int *mattuples3, uint64_t row_offset, uint64_t col_offset, int ckthr, char **dfd_col_seq_gpu, char **dfd_row_seq_gpu, uint64_t* align_cnts);
 
 void test1();
